pyproject.toml
[project]
name = "bookmark-analytics-toolkit"
version = "0.1.0"
description = "Web browser bookmark analysis application with GUI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "polars>=0.20.0",
    "matplotlib>=3.8.0",
    "plotly>=5.18.0",
    "seaborn>=0.13.0",
    "matplotlib-fontja>=1.1.0",
    "streamlit>=1.51.0",
    "sudachipy>=0.6.0",
    "sudachidict-full>=20240716",
    "wordcloud>=1.9.0",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "ruff>=0.1.0",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bookmark_analytics_toolkit"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "N", "W"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[dependency-groups]
dev = [
    "pip-licenses>=5.5.0",
]
